% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismayFizzBuzz.R
\name{fizz_buzz}
\alias{fizz_buzz}
\title{FizzBuzz Game}
\usage{
fizz_buzz(x)
}
\arguments{
\item{x}{A positive integer}
}
\value{
"Fizz", "Buzz", "FizzBuzz", or the value of `x` based on description
}
\description{
If x is a positive integer (i.e., greater than 0) and:
divisible by 3, return the string 'Fizz'
divisible by 5, return the string 'Buzz'
divisible by both 3 and 5, return the string 'FizzBuzz'.
If x doesn't satisfy any of the above conditions, simply return x.
}
\examples{

# divisible by 3, output "Fizz"
fizz_buzz(21L)

# divisible by 5, output "Buzz"
fizz_buzz(35L)

# divisible by 3 and 5, output "FizzBuzz"
fizz_buzz(60L)

# divisible neither 3 nor 5, output 41
fizz_buzz(41L)


}
